new: Fix bounds rendering on UI, add numbers to multi-field
