+++ calibrate/autolayout.py
@@ -110,74 +110,3 @@ LAYOUTS = {"STANDARD": Layout("Standard", (0.5,0.5), PREVIEW_LAYOUTS["STANDARD"]
            "MOC_BOTRIGHT": Layout("MaxoutClub", (0.556,0.669), PREVIEW_LAYOUTS["MOC4pRight"]) #ctwc 4p
           }
 
-class Rect:
-    def __init__(self, left, top, right, bottom):
-        self.left = left
-        self.top = top
-        self.right = right
-        self.bottom = bottom
-    
-    @property
-    def width(self):
-        return self.right - self.left
-    
-    @property
-    def height(self):
-        return self.bottom - self.top
-
-    @property
-    def area(self):
-        return self.width * self.height
-
-    @property
-    def centre(self):
-        return (self.left + 0.5* self.width,
-               self.top + 0.5 * self.height)
-
-    def to_array(self):
-        return (self.left, self.top, self.right, self.bottom)
-    
-    def __str__(self):
-        return str(self.to_array())
-   
-    def __eq__(self, other):
-        if isinstance(other, Rect):
-            return (self.left == other.left and
-                   self.top == other.top and
-                   self.right == other.right and
-                   self.bottom == other.bottom)
-        return False
-    
-    def contains(self, xy):
-        return (self.left <= xy[0] <= self.right and 
-               self.top <= xy[1] <= self.bottom)
-
-    def within(self, yx):
-        """
-        returns if rectangle is within an image with given y/x size.
-        """
-        return (0 <= self.left <= self.right <= yx[1] and 
-                0 <= self.top <= self.bottom <= yx[0])
-           
-    def multiply(self, constant):
-        self.left = self.left * constant
-        self.top = self.top * constant
-        self.right = self.right * constant
-        self.bottom = self.bottom* constant
-    
-    def round_to_int(self):
-        self.left = round(self.left)
-        self.top = round(self.top)
-        self.right = round(self.right)
-        self.bottom= round(self.bottom)
-
-    def sq_distance(self, point):
-        """
-        square distance from center of rect to point.
-        Because math.sqrt is expensive yo
-        """
-        you = self.centre #you self.centred b****
-        xdist = (you[0] - point[0]) 
-        ydist = (you[1] - point[1])
-        result = xdist*xdist + ydist*ydist
-        return result
